Exploring possibility of moving DagAttributeTypes to execution API spec
