+++ task-sdk/src/airflow/sdk/api/datamodels/_generated.py
@@ -485,6 +485,11 @@ class TriggerRule(str, Enum):
     ALL_SKIPPED = "all_skipped"
 
 
+class DagAttributeTypes(str, Enum):
+    OPERATOR = "operator"
+    TASKGROUP = "taskgroup"
+
+
 class AssetReferenceAssetEventDagRun(BaseModel):
     """
     Schema for AssetModel used in AssetEventDagRunReference.
